Even more coverance, just refactor to check if affects
