+++ gamey/src/bot_server/play.rs
@@ -10,9 +10,9 @@ use axum::extract::Query;
 #[derive(Deserialize)]
 pub struct PlayParams {
     /// The API version (e.g., "v1").
-    api_version: String,
+    pub api_version: String,
     /// The identifier of the bot to use for move selection.
-    bot_id: String,
+    pub bot_id: String,
 }
 
 /// Response returned by the play endpoint on success.
@@ -183,8 +183,8 @@ pub async fn play(
 
 #[derive(Deserialize)]
 pub struct PlayQuery {
-    position: String,
-    bot_id: Option<String>,
+    pub position: String,
+    pub bot_id: Option<String>,
 }
 
 /// GET handler at `/play?bot_id=<id>&position=<url-encoded JSON>`.
@@ -201,857 +201,4 @@ pub async fn play_get(
     })?;
     let coords = play(query.bot_id.as_deref(), Json(yen)).await?;
     Ok(Json(coords))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_play_response_creation() {
-        let yen = YEN::new(3, 1, vec!['B', 'R'], "B/BR/.R.".to_string());
-        let response = PlayResponse {
-            api_version: "v1".to_string(),
-            bot_id: "minimax_bot".to_string(),
-            coords: crate::Coordinates::new(0, 0, 2),
-            position: yen,
-            game_over: false,
-            winner: None,
-        };
-        assert_eq!(response.api_version, "v1");
-        assert_eq!(response.bot_id, "minimax_bot");
-        assert!(!response.game_over);
-        assert!(response.winner.is_none());
-    }
-
-    #[test]
-    fn test_play_response_with_winner() {
-        let yen = YEN::new(3, 1, vec!['B', 'R'], "B/BB/BBB".to_string());
-        let response = PlayResponse {
-            api_version: "v1".to_string(),
-            bot_id: "minimax_bot".to_string(),
-            coords: crate::Coordinates::new(0, 0, 2),
-            position: yen,
-            game_over: true,
-            winner: Some("B".to_string()),
-        };
-        assert!(response.game_over);
-        assert_eq!(response.winner, Some("B".to_string()));
-    }
-
-    #[test]
-    fn test_play_response_serialize() {
-        let yen = YEN::new(3, 0, vec!['B', 'R'], "./../.".to_string());
-        let response = PlayResponse {
-            api_version: "v1".to_string(),
-            bot_id: "random_bot".to_string(),
-            coords: crate::Coordinates::new(1, 0, 1),
-            position: yen,
-            game_over: false,
-            winner: None,
-        };
-        let json = serde_json::to_string(&response).unwrap();
-        assert!(json.contains("\"api_version\":\"v1\""));
-        assert!(json.contains("\"bot_id\":\"random_bot\""));
-        assert!(json.contains("\"game_over\":false"));
-        assert!(json.contains("\"winner\":null"));
-    }
-
-    #[test]
-    fn test_play_response_deserialize() {
-        let json = r#"{
-            "api_version": "v1",
-            "bot_id": "random_bot",
-            "coords": {"x":1,"y":0,"z":1},
-            "position": {"size":3,"turn":1,"players":["B","R"],"layout":"./../."},
-            "game_over": false,
-            "winner": null
-        }"#;
-        let response: PlayResponse = serde_json::from_str(json).unwrap();
-        assert_eq!(response.api_version, "v1");
-        assert_eq!(response.bot_id, "random_bot");
-        assert!(!response.game_over);
-    }
-
-    /// Picks any bot name available in the default state, so tests don't
-    /// hard-code a specific bot id that might not exist in this build.
-    fn any_available_bot_id() -> String {
-        let state = create_default_state();
-        state
-            .bots()
-            .names()
-            .first()
-            .cloned()
-            .expect("default state should expose at least one bot")
-    }
-
-    /// Empty 3-row Y board: row 1 has 1 cell, row 2 has 2 cells, row 3 has 3 cells.
-    fn empty_board_size_3() -> String {
-        "./../...".to_string()
-    }
-
-    #[tokio::test]
-    async fn test_play_returns_coords_on_valid_position() {
-        let bot_id = any_available_bot_id();
-        let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-        let result = play(Option::from(&*bot_id), Json(yen)).await;
-
-        assert!(
-            result.is_ok(),
-            "play should succeed on a valid empty position, got: {:?}",
-            result.err().map(|Json(e)| e)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_play_unknown_bot_returns_error() {
-        let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-        let result = play(Option::from("__definitely_not_a_real_bot__"), Json(yen)).await;
-
-        let Json(err) = result.expect_err("unknown bot should produce an error");
-        let msg = format!("{:?}", err);
-        assert!(
-            msg.contains("Bot not found"),
-            "error should mention 'Bot not found', got: {}",
-            msg
-        );
-        assert!(
-            msg.contains("__definitely_not_a_real_bot__"),
-            "error should echo the bad bot id, got: {}",
-            msg
-        );
-    }
-
-    #[tokio::test]
-    async fn test_play_invalid_yen_returns_error() {
-        let bot_id = any_available_bot_id();
-        // Layout that does not match the declared size -> GameY::try_from must fail.
-        let yen = YEN::new(3, 0, vec!['B', 'R'], "garbage-layout".to_string());
-
-        let result = play(Option::from(&*bot_id), Json(yen)).await;
-
-        let Json(err) = result.expect_err("invalid YEN should produce an error");
-        let msg = format!("{:?}", err);
-        assert!(
-            msg.contains("Invalid YEN position"),
-            "error should mention 'Invalid YEN position', got: {}",
-            msg
-        );
-    }
-
-    #[tokio::test]
-    async fn test_play_respects_yen_turn() {
-        // Same empty board, but turn = 1: play must still return coords.
-        let bot_id = any_available_bot_id();
-        let yen = YEN::new(3, 1, vec!['B', 'R'], empty_board_size_3());
-
-        let result = play(Option::from(&*bot_id), Json(yen)).await;
-
-        assert!(
-            result.is_ok(),
-            "play should honor a non-zero turn on a non-terminal position, got: {:?}",
-            result.err().map(|Json(e)| e)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_play_no_bot() {
-        let yen = YEN::new(3, 1, vec!['B', 'R'], empty_board_size_3());
-
-        let result = play(None, Json(yen)).await;
-
-        assert!(
-            result.is_ok(),
-            "play should honor a minimax bot, got: {:?}",
-            result.err().map(|Json(e)| e)
-        );
-        fn create_test_state() -> AppState {
-            create_default_state()
-        }
-
-        #[tokio::test]
-        async fn test_player_play_success_on_valid_position() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: bot_id.clone(),
-            };
-
-            let result = player_play(
-                State(state),
-                Path(params),
-                Json(yen),
-            ).await;
-
-            assert!(
-                result.is_ok(),
-                "player_play should succeed on a valid empty position, got: {:?}",
-                result.err()
-            );
-
-            let Json(response) = result.unwrap();
-            assert_eq!(response.api_version, "v1");
-            assert_eq!(response.bot_id, bot_id);
-            assert!(!response.game_over);
-            assert!(response.winner.is_none());
-        }
-
-        #[tokio::test]
-        async fn test_player_play_invalid_api_version() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v999".to_string(), // Invalid version
-                bot_id: bot_id.clone(),
-            };
-
-            let result = player_play(
-                State(state),
-                Path(params),
-                Json(yen),
-            ).await;
-
-            assert!(
-                result.is_err(),
-                "player_play should reject invalid API version"
-            );
-        }
-
-        #[tokio::test]
-        async fn test_player_play_unknown_bot() {
-            let state = create_test_state();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: "__unknown_bot_xyz__".to_string(),
-            };
-
-            let result = player_play(
-                State(state),
-                Path(params),
-                Json(yen),
-            ).await;
-
-            assert!(
-                result.is_err(),
-                "player_play should return error for unknown bot"
-            );
-
-            let Json(err) = result.unwrap_err();
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Bot not found"),
-                "error should mention 'Bot not found', got: {}",
-                msg
-            );
-        }
-
-        #[tokio::test]
-        async fn test_player_play_invalid_yen_position() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], "invalid-layout".to_string());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id,
-            };
-
-            let result = player_play(
-                State(state),
-                Path(params),
-                Json(yen),
-            ).await;
-
-            assert!(
-                result.is_err(),
-                "player_play should reject invalid YEN position"
-            );
-
-            let Json(err) = result.unwrap_err();
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Invalid YEN position"),
-                "error should mention 'Invalid YEN position', got: {}",
-                msg
-            );
-        }
-
-        #[tokio::test]
-        async fn test_player_play_respects_turn() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 1, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id,
-            };
-
-            let result = player_play(
-                State(state),
-                Path(params),
-                Json(yen),
-            ).await;
-
-            assert!(
-                result.is_ok(),
-                "player_play should honor non-zero turn, got: {:?}",
-                result.err()
-            );
-
-            let Json(response) = result.unwrap();
-            assert_eq!(response.api_version, "v1");
-        }
-
-        #[tokio::test]
-        async fn test_player_play_returns_valid_response_structure() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: bot_id.clone(),
-            };
-
-            let result = player_play(
-                State(state),
-                Path(params),
-                Json(yen),
-            ).await;
-
-            let Json(response) = result.unwrap();
-
-            // Verify all fields are properly set
-            assert_eq!(response.api_version, "v1");
-            assert_eq!(response.bot_id, bot_id);
-            assert!(!response.game_over); // Board is not finished
-            assert!(response.winner.is_none()); // No winner yet
-            // coords should be valid (from the successful move)
-            // position should reflect the updated board
-        }
-
-        // ==================== NEW TESTS: play_get ====================
-
-        #[tokio::test]
-        async fn test_play_get_success_with_valid_position() {
-            let position_json = r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(
-                result.is_ok(),
-                "play_get should succeed with valid position, got: {:?}",
-                result.err()
-            );
-        }
-
-        #[tokio::test]
-        async fn test_play_get_invalid_position_json() {
-            let query = PlayQuery {
-                position: "{ invalid json }".to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(
-                result.is_err(),
-                "play_get should fail on invalid JSON"
-            );
-
-            let Json(err) = result.unwrap_err();
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Invalid position query parameter"),
-                "error should mention JSON parse error, got: {}",
-                msg
-            );
-        }
-
-        #[tokio::test]
-        async fn test_play_get_missing_bot_id_uses_default() {
-            let position_json = r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: None, // Should default to "minimax_bot"
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(
-                result.is_ok(),
-                "play_get should succeed with default bot, got: {:?}",
-                result.err()
-            );
-        }
-
-        #[tokio::test]
-        async fn test_play_get_unknown_bot() {
-            let position_json = r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some("__nonexistent_bot__".to_string()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(
-                result.is_err(),
-                "play_get should fail with unknown bot"
-            );
-
-            let Json(err) = result.unwrap_err();
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Bot not found"),
-                "error should mention 'Bot not found', got: {}",
-                msg
-            );
-        }
-
-        #[tokio::test]
-        async fn test_play_get_invalid_yen_in_position() {
-            let position_json = r#"{"size":3,"turn":0,"players":["B","R"],"layout":"invalid"}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(
-                result.is_err(),
-                "play_get should fail on invalid YEN layout"
-            );
-
-            let Json(err) = result.unwrap_err();
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Invalid YEN position"),
-                "error should mention 'Invalid YEN position', got: {}",
-                msg
-            );
-        }
-
-        #[tokio::test]
-        async fn test_play_get_returns_coordinates() {
-            let position_json = r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            let Json(coords) = result.unwrap();
-            // Coordinates should be valid (not null, within board bounds)
-            // This is a basic check; specific validation depends on Coordinates structure
-            assert!(true); // Placeholder for coordinate validation
-        }
-
-        #[tokio::test]
-        async fn test_play_get_with_multiple_positions() {
-            let bot_id = any_available_bot_id();
-
-            // Test 1: Empty board
-            let query1 = PlayQuery {
-                position: r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#.to_string(),
-                bot_id: Some(bot_id.clone()),
-            };
-            assert!(play_get(Query(query1)).await.is_ok());
-
-            // Test 2: Board with one move
-            let query2 = PlayQuery {
-                position: r#"{"size":3,"turn":1,"players":["B","R"],"layout":"B/../..."}"#.to_string(),
-                bot_id: Some(bot_id.clone()),
-            };
-            assert!(play_get(Query(query2)).await.is_ok());
-        }
-
-        #[tokio::test]
-        async fn test_play_get_malformed_json_structure() {
-            let query = PlayQuery {
-                position: r#"{"size":"three","players":["B","R"]}"#.to_string(), // Missing layout and turn
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(result.is_err(), "play_get should fail on incomplete JSON structure");
-        }
-
-        #[tokio::test]
-        async fn test_play_get_url_encoded_position() {
-            // Simulating a URL-encoded position (should still parse)
-            let position_json = r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let result = play_get(Query(query)).await;
-
-            assert!(
-                result.is_ok(),
-                "play_get should handle URL-encoded positions, got: {:?}",
-                result.err()
-            );
-        }
-
-        // ==================== ADDITIONAL COVERAGE: player_play ====================
-
-        /// Covers the `game.next_player() == None` branch: when the YEN already
-        /// describes a finished game, player_play must return the
-        /// "Game is already over" error before asking the bot for a move.
-        #[tokio::test]
-        async fn test_player_play_game_already_over() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            // A fully-filled board where B has a winning chain.
-            // Layout for size 3: row1=1 cell, row2=2 cells, row3=3 cells.
-            let yen = YEN::new(3, 0, vec!['B', 'R'], "B/BR/BRR".to_string());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id,
-            };
-
-            let result = player_play(State(state), Path(params), Json(yen)).await;
-
-            // If the position parses as terminal, we expect the "already over" error.
-            // If for any reason the engine considers it ongoing, the call should at
-            // least not panic — but in the standard Y rules a full board is terminal.
-            if let Err(Json(err)) = result {
-                let msg = format!("{:?}", err);
-                assert!(
-                    msg.contains("Game is already over")
-                        || msg.contains("Invalid YEN position"),
-                    "expected 'Game is already over' or 'Invalid YEN position', got: {}",
-                    msg
-                );
-            }
-        }
-
-        /// Covers the successful path where `game_over == true` and the winner
-        /// is reported. We don't assert which color wins (depends on engine
-        /// rules) — only that, when game_over is true, winner is Some(...).
-        #[tokio::test]
-        async fn test_player_play_reports_winner_when_game_over() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            // Near-terminal position: one empty cell left. The bot is forced
-            // to play it, which should end the game on this move.
-            // Trying a layout where B only needs one more move on a size-3 board.
-            let yen = YEN::new(3, 0, vec!['B', 'R'], "B/BR/.RB".to_string());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: bot_id.clone(),
-            };
-
-            let result = player_play(State(state), Path(params), Json(yen)).await;
-
-            if let Ok(Json(response)) = result {
-                // If the move ended the game, winner must be set and be "B" or "R".
-                if response.game_over {
-                    let w = response.winner.as_deref();
-                    assert!(
-                        w == Some("B") || w == Some("R"),
-                        "winner should be 'B' or 'R' when game_over=true, got: {:?}",
-                        response.winner
-                    );
-                } else {
-                    // Otherwise the game continues and winner stays None.
-                    assert!(response.winner.is_none());
-                }
-                assert_eq!(response.api_version, "v1");
-                assert_eq!(response.bot_id, bot_id);
-            }
-        }
-
-        /// Covers the successful path and asserts the response actually carries
-        /// a meaningful YEN position back to the caller (not just metadata).
-        #[tokio::test]
-        async fn test_player_play_response_contains_updated_position() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: bot_id.clone(),
-            };
-
-            let result = player_play(State(state), Path(params), Json(yen)).await;
-            let Json(response) = result.expect("valid empty board should succeed");
-
-            // The returned position must round-trip back into a GameY: it
-            // represents the board *after* the bot's move.
-            let parsed = GameY::try_from(response.position.clone());
-            assert!(
-                parsed.is_ok(),
-                "response.position should be a valid YEN, got error: {:?}",
-                parsed.err()
-            );
-
-            // After exactly one move on an empty 3-row board, the turn
-            // counter in the returned YEN should have advanced.
-            assert!(
-                response.position.turn() >= 1,
-                "turn counter should advance after a move, got turn={}",
-                response.position.turn()
-            );
-        }
-
-        /// Covers invalid-API-version error: also asserts the error payload
-        /// echoes back the bot_id and api_version provided by the caller, so
-        /// future regressions in error construction get caught.
-        #[tokio::test]
-        async fn test_player_play_invalid_api_version_error_payload() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v999".to_string(),
-                bot_id: bot_id.clone(),
-            };
-
-            let Json(err) = player_play(State(state), Path(params), Json(yen))
-                .await
-                .expect_err("invalid api version must error");
-
-            let msg = format!("{:?}", err);
-            // We don't pin the exact wording, but the error string must be
-            // non-empty and reference the version somehow.
-            assert!(!msg.is_empty(), "error message should not be empty");
-        }
-
-        /// Covers the unknown-bot branch and asserts the error message lists
-        /// the available bots — a behavior promised by the source code.
-        #[tokio::test]
-        async fn test_player_play_unknown_bot_lists_available() {
-            let state = create_test_state();
-            let yen = YEN::new(3, 0, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: "__totally_made_up__".to_string(),
-            };
-
-            let Json(err) = player_play(State(state), Path(params), Json(yen))
-                .await
-                .expect_err("unknown bot must error");
-
-            let msg = format!("{:?}", err);
-            assert!(msg.contains("Bot not found"), "got: {}", msg);
-            assert!(
-                msg.contains("available bots"),
-                "error should list available bots, got: {}",
-                msg
-            );
-            assert!(
-                msg.contains("__totally_made_up__"),
-                "error should echo the bad bot id, got: {}",
-                msg
-            );
-        }
-
-        /// Covers `force_turn` being honored: same empty board but with
-        /// turn=1, the response's bot_id and api_version must still match
-        /// what the caller supplied (regression test for parameter routing).
-        #[tokio::test]
-        async fn test_player_play_preserves_request_metadata() {
-            let state = create_test_state();
-            let bot_id = any_available_bot_id();
-            let yen = YEN::new(3, 1, vec!['B', 'R'], empty_board_size_3());
-
-            let params = PlayParams {
-                api_version: "v1".to_string(),
-                bot_id: bot_id.clone(),
-            };
-
-            let Json(response) = player_play(State(state), Path(params), Json(yen))
-                .await
-                .expect("non-zero turn on empty board should succeed");
-
-            assert_eq!(response.api_version, "v1");
-            assert_eq!(response.bot_id, bot_id);
-        }
-
-        // ==================== ADDITIONAL COVERAGE: play_get ====================
-
-        /// The existing `test_play_get_returns_coordinates` only does
-        /// `assert!(true)` — this test actually inspects the returned
-        /// coordinates to make sure the handler produces a structurally
-        /// valid result.
-        #[tokio::test]
-        async fn test_play_get_returns_in_bounds_coordinates() {
-            let position_json =
-                r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let Json(coords) = play_get(Query(query))
-                .await
-                .expect("valid position must yield coordinates");
-
-            // The chosen coordinates must serialize cleanly — i.e. they're a
-            // real Coordinates value, not some default/garbage.
-            let serialized = serde_json::to_string(&coords)
-                .expect("coordinates must be serializable");
-            assert!(
-                !serialized.is_empty() && serialized != "null",
-                "coordinates payload should be non-empty, got: {}",
-                serialized
-            );
-        }
-
-        /// Covers the default-bot path explicitly: when `bot_id` is None we
-        /// fall back to "minimax_bot". If that bot ever disappears from the
-        /// default state, this test will start failing — which is exactly
-        /// what we want, because production code hard-codes that name.
-        #[tokio::test]
-        async fn test_play_get_default_bot_is_minimax() {
-            let state = create_default_state();
-            let names = state.bots().names();
-            let has_minimax = names.iter().any(|n| n == "minimax_bot");
-
-            // Only meaningful if the default state actually exposes minimax_bot.
-            // If it doesn't, the production default itself is broken — assert that.
-            assert!(
-                has_minimax,
-                "default state should expose 'minimax_bot' as the fallback bot, got: {:?}",
-                names
-            );
-
-            let position_json =
-                r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: None,
-            };
-
-            let result = play_get(Query(query)).await;
-            assert!(
-                result.is_ok(),
-                "play_get with no bot_id should fall back to minimax_bot, got: {:?}",
-                result.err()
-            );
-        }
-
-        /// Covers the unknown-bot error path through `play_get` and asserts
-        /// the error lists available bots (forwarded from `play`).
-        #[tokio::test]
-        async fn test_play_get_unknown_bot_lists_available() {
-            let position_json =
-                r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-            let query = PlayQuery {
-                position: position_json.to_string(),
-                bot_id: Some("__nope__".to_string()),
-            };
-
-            let Json(err) = play_get(Query(query))
-                .await
-                .expect_err("unknown bot must error");
-
-            let msg = format!("{:?}", err);
-            assert!(msg.contains("Bot not found"), "got: {}", msg);
-            assert!(
-                msg.contains("available bots"),
-                "error should list available bots, got: {}",
-                msg
-            );
-        }
-
-        /// Covers the empty-string position branch: serde_json should reject
-        /// empty input, producing the "Invalid position query parameter" error.
-        #[tokio::test]
-        async fn test_play_get_empty_position_string() {
-            let query = PlayQuery {
-                position: String::new(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let Json(err) = play_get(Query(query))
-                .await
-                .expect_err("empty position must error");
-
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Invalid position query parameter"),
-                "got: {}",
-                msg
-            );
-        }
-
-        /// Covers all available bots (not just the first one). If the project
-        /// exposes several bots, this exercises each one through play_get,
-        /// catching regressions where one specific bot misbehaves on an
-        /// empty board.
-        #[tokio::test]
-        async fn test_play_get_works_for_every_available_bot() {
-            let state = create_default_state();
-            let names = state.bots().names();
-            assert!(!names.is_empty(), "default state should expose >=1 bot");
-
-            let position_json =
-                r#"{"size":3,"turn":0,"players":["B","R"],"layout":"./../..."}"#;
-
-            for bot_id in names {
-                let query = PlayQuery {
-                    position: position_json.to_string(),
-                    bot_id: Some(bot_id.clone()),
-                };
-
-                let result = play_get(Query(query)).await;
-                assert!(
-                    result.is_ok(),
-                    "play_get should succeed for bot '{}', got: {:?}",
-                    bot_id,
-                    result.err()
-                );
-            }
-        }
-
-        /// Covers a position where the size declared in YEN doesn't match the
-        /// layout — should be rejected at GameY::try_from with the
-        /// "Invalid YEN position" error, going through `play` -> `play_get`.
-        #[tokio::test]
-        async fn test_play_get_size_layout_mismatch() {
-            // size=3 but layout only has data for one row.
-            let query = PlayQuery {
-                position: r#"{"size":3,"turn":0,"players":["B","R"],"layout":"."}"#
-                    .to_string(),
-                bot_id: Some(any_available_bot_id()),
-            };
-
-            let Json(err) = play_get(Query(query))
-                .await
-                .expect_err("size/layout mismatch must error");
-
-            let msg = format!("{:?}", err);
-            assert!(
-                msg.contains("Invalid YEN position"),
-                "got: {}",
-                msg
-            );
-        }
-    }
 }
